Relocate task tests from tests/python/pants_test/task/ to the appropriate backend
