+++ tests/python/pants_test/backend/codegen/tasks/test_protobuf_integration.py
