+++ tests/python/pants_test/backend/jvm/tasks/test_check_published_deps.py
