+++ tests/python/pants_test/backend/jvm/tasks/test_jar_publish_integration.py
