+++ tests/python/pants_test/backend/jvm/tasks/test_junit_tests_integration.py
