+++ tests/python/pants_test/backend/jvm/tasks/test_jvm_bundle_integration.py
