+++ tests/python/pants_test/tasks/BUILD
@@ -17,15 +17,9 @@ target(
   name = 'tasks',
   dependencies = [
     ':builddict',
-    ':check_published_deps',
     ':detect_duplicates',
     ':execution_graph',
     ':group_task',
-    ':jar_create',
-    ':jar_publish',
-    ':jar_task',
-    ':jvm_task',
-    ':jvmdoc_gen',
     ':list_goals',
     ':list_owners',
     ':listtargets',
@@ -41,14 +35,8 @@ target(
 target(
   name = 'integration',
   dependencies = [
-    ':antlr_integration',
     ':bootstrap_jvm_tools_integration',
     ':ivy_resolve_integration',
-    ':jar_publish_integration',
-    ':junit_tests_integration',
-    ':jvm_bundle_integration',
-    ':jvm_run_integration',
-    ':protobuf_integration',
     ':scalastyle_integration',
   ],
 )
@@ -60,16 +48,6 @@ python_tests(
     'tests/python/pants_test:int-test',
     'src/python/pants/util:contextutil',
   ]
-)
-
-python_tests(
-  name = 'antlr_integration',
-  sources = ['test_antlr_integration.py'],
-  dependencies = [
-    'src/python/pants/fs',
-    'src/python/pants/util:contextutil',
-    'tests/python/pants_test:int-test',
-    ]
 )
 
 python_tests(
@@ -98,33 +76,6 @@ python_tests(
 )
 
 python_tests(
-  name = 'junit_tests_integration',
-  sources = ['test_junit_tests_integration.py'],
-  dependencies = [
-    'src/python/pants/util:contextutil',
-    'tests/python/pants_test:int-test',
-  ],
-)
-
-python_tests(
-  name = 'jvm_bundle_integration',
-  sources = ['test_jvm_bundle_integration.py'],
-  dependencies = [
-    'src/python/pants/fs',
-    'src/python/pants/util:contextutil',
-    'tests/python/pants_test:int-test',
-  ],
-)
-
-python_tests(
-  name = 'jvm_run_integration',
-  sources = ['test_jvm_run_integration.py'],
-  dependencies = [
-    'tests/python/pants_test:int-test',
-  ],
-)
-
-python_tests(
   name = 'bootstrap_jvm_tools_integration',
   sources = ['test_bootstrap_jvm_tools_integration.py'],
   dependencies = [
@@ -142,21 +93,6 @@ python_tests(
     'src/python/pants/backend/jvm:plugin',
     'src/python/pants/backend/python:plugin',
     'tests/python/pants_test:base_test',
-  ]
-)
-
-python_tests(
-  name = 'check_published_deps',
-  sources = ['test_check_published_deps.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/core/targets:common',
-    'src/python/pants/backend/jvm:artifact',
-    'src/python/pants/backend/jvm:repository',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/tasks:check_published_deps',
-    'src/python/pants/build_graph',
   ]
 )
 
@@ -204,58 +140,6 @@ python_tests(
 )
 
 python_tests(
-  name = 'jar_create',
-  sources = ['test_jar_create.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/android/targets:android',
-    'src/python/pants/backend/codegen/targets:java',
-    'src/python/pants/backend/core/targets:common',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/targets:scala',
-    'src/python/pants/backend/jvm/tasks:jar_create',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/build_graph',
-    'src/python/pants/java/jar:manifest',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'tests/python/pants_test/base:context_utils',
-    'tests/python/pants_test/jvm:jar_task_test_base',
-  ],
-)
-
-python_tests(
-  name = 'jar_publish',
-  sources = ['test_jar_publish.py'],
-  dependencies = [
-    '3rdparty/python:mock',
-    ':task_test_base',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/tasks:jar_publish',
-    'src/python/pants/base:generator',
-    'src/python/pants/base:source_root',
-    'src/python/pants/build_graph',
-    'src/python/pants/scm:scm',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'tests/python/pants_test/testutils',
-  ],
-)
-
-python_tests(
-  name = 'jar_publish_integration',
-  sources = ['test_jar_publish_integration.py'],
-  dependencies = [
-    'src/python/pants/base:build_environment',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'tests/python/pants_test:int-test',
-  ],
-)
-
-python_tests(
   name = 'list_goals',
   sources = ['test_list_goals.py'],
   dependencies = [
@@ -288,14 +172,6 @@ python_tests(
     'src/python/pants/backend/core/tasks:minimal_cover',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
-  name = 'protobuf_integration',
-  sources = ['test_protobuf_integration.py'],
-  dependencies = [
-    'tests/python/pants_test:int-test',
   ],
 )
 
@@ -362,23 +238,6 @@ python_tests(
 )
 
 python_tests(
-  name = 'jar_task',
-  sources = ['test_jar_task.py'],
-  dependencies = [
-    '3rdparty/python/twitter/commons:twitter.common.collections',
-    '3rdparty/python:six',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/tasks:jar_task',
-    'src/python/pants/base:source_root',
-    'src/python/pants/build_graph',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'tests/python/pants_test/jvm:jvm_tool_task_test_base',
-  ]
-)
-
-python_tests(
   name = 'group_task',
   sources = ['test_group_task.py'],
   dependencies = [
@@ -387,27 +246,6 @@ python_tests(
     'src/python/pants/build_graph',
     'src/python/pants/engine',
     'tests/python/pants_test:base_test',
-  ]
-)
-
-python_tests(
-  name = 'jvm_task',
-  sources = ['test_jvm_task.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/jvm/tasks:classpath_products',
-    'src/python/pants/backend/jvm/tasks:jvm_task',
-    'tests/python/pants_test/jvm:jvm_task_test_base'
-  ]
-)
-
-python_tests(
-  name = 'jvmdoc_gen',
-  sources = ['test_jvmdoc_gen.py'],
-  dependencies = [
-    'src/python/pants/backend/jvm/tasks:jvmdoc_gen',
-    'src/python/pants/base:exceptions',
-    'tests/python/pants_test/jvm:jvm_task_test_base'
   ]
 )
